Refactor E2E tests to use shared test-helpers.ts
- Remove duplicate exec() functions from all 19 test files
- Import exec variants from centralized test-helpers.ts
- exec() for standard tests, execWithFilter for commit-commands,
  execProduction for docker-commands
- All tests use isolated environment (clears PRLT_HQ_PATH, etc.)
